+++ resources/onnx/layer-zoo/onnx-layer-zoo.py
@@ -135,6 +135,16 @@ def squeeze_node_with_axes():
     )
     return ("squeeze_with_axes", node, [1,2,1,3], [2,3], [axes])
 
+def unsqueeze_node():
+    axes = make_constant_int_node( "axes", [-2, 0] )
+
+    node = onnx.helper.make_node(
+        "Unsqueeze",
+        inputs=[input_name, "axes"],
+        outputs=[output_name]
+    )
+    return ("unsqueeze", node, [2,3], [1,1,2,3], [axes])
+
 def batch_normalization_node():
     scale = make_constant_float_node("scale", [0.5, 1, 2])
     bias = make_constant_float_node("bias", [0, 1, 0])
@@ -299,6 +309,7 @@ if __name__ == "__main__":
     make_network(*transpose_node())
     make_network(*squeeze_node())
     make_network(*squeeze_node_with_axes())
+    make_network(*unsqueeze_node())
     make_network(*batch_normalization_node())
     make_network(*max_pool_node())
     make_network(*conv_node())