+++ src/input_parsers/tests/Test_OnnxParser.h
@@ -188,6 +188,13 @@ public:
         run_test( "squeeze_with_axes", input, output );
     }
 
+    void test_unsqueeze()
+    {
+        Vector<double> input = { 0, 1, -2, 0.5, 1.5, -3 };
+        Vector<double> output = { 0, 1, -2, 0.5, 1.5, -3 };
+        run_test( "unsqueeze", input, output );
+    }
+
     // Reference implementation for batch normalisation
     double batchnorm_fn( double input,
                          double epsilon,